requirements.txt
pyyaml==6.0.1
streamlit==1.35.0
st-gsheets-connection==0.0.4